...WhyGod.png
I fixed the enemies + player being too small by splicing the sprites for the tiles and walls in Photoshop but it has made MORE PROBLEMS than what it's FIXED.
EVERYTHING IS TINY DEARGODWHY.

+++ Roguelite/Assets/Scripts/LevelGenerator.cs
@@ -23,7 +23,7 @@ public class LevelGenerator : MonoBehaviour
 	public int tileAmount;
 
 	//the space between each movement of the generator- which is going to be the size of the tiles
-	public int tileSize;
+	public float tileSize;
